test(textlint-kernel): Convert proccesor-helper.ts
